+++ app/src/main/java/pt/ulusofona/deisi/a2020/cm/g2/ui/listeners/OnBatteryCurrentListener.kt
@@ -2,5 +2,5 @@ package pt.ulusofona.deisi.a2020.cm.g2.ui.listeners
 
 interface OnBatteryCurrentListener {
 
-    fun onCurrentChanged(current: Int)
+    fun onCurrentChanged(current: Float)
 }
